Move interactive test to scripts, minor reorg
